organized files and renamed camer.js reducer to meal.js
