+++ foodprintReactNative/reducers/index.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux';
 const rootReducer = combineReducers({
   welcome: require('./default').default,
   //cameraResults: require('./camera').default,
-  meal: require('./camera').default,
+  meal: require('./meal').default,
   auth: require('./auth').default
 });
